+++ SwiftUIDemoAppUITests/Tests/DeeplinkTests.swift
@@ -1,8 +1,5 @@
 //
-//  DeeplinkTests.swift
-//  SwiftUIDemoAppUITests
-//
-//  Created by Ilias Pavlidakis on 25/9/23.
+// Copyright © 2023 Stream.io Inc. All rights reserved.
 //
 
 import XCTest
@@ -60,88 +57,5 @@ final class DeeplinkTests: StreamTestCase {
                 .assertCallControls()
                 .assertParticipantsAreVisible(count: 1)
         }
-    }
-}
-
-extension XCUIDevice {
-
-    fileprivate func open(_ url: URL) {
-        DispatchQueue.main.async {
-            // Need to async this because XCUIDevice.shared.system.open
-            // synchronously waits for a button to be pressed the first time.
-            self.system.open(url)
-        }
-    }
-}
-
-enum Safari {
-
-    private static let safari = XCUIApplication(bundleIdentifier: "com.apple.mobilesafari")
-
-    fileprivate static func go(to url: URL) {
-        safari.textFields["Address"].tap()
-        safari.typeText(url.absoluteString)
-        let goButton = safari.buttons["Go"]
-        if goButton.waitForExistence(timeout: 5) {
-            safari.tapKeyboardKey("Go")
-        }
-    }
-
-    fileprivate static func open(_ url: URL) {
-        safari.launch()
-
-        _ = safari.wait(for: .runningForeground, timeout: 5)
-
-        // Type the deeplink and execute it
-        let firstLaunchContinueButton = safari.buttons["Continue"]
-        if firstLaunchContinueButton.exists {
-            firstLaunchContinueButton.tap()
-        }
-
-        go(to: url)
-    }
-
-    fileprivate static func openApp(_ url: URL) {
-        open(url)
-
-        let openButton = safari.buttons["Open"]
-        if openButton.waitForExistence(timeout: 5) {
-            openButton.tap()
-        }
-    }
-
-    fileprivate static func openUniversalLinkFromSmartBanner(_ url: URL) {
-        open(url)
-
-        let allowButton1 = safari.alerts.buttons["Allow"]
-        if allowButton1.waitForExistence(timeout: 5) {
-            allowButton1.tap()
-            let allowButton2 = safari.alerts.buttons["Allow"]
-            if allowButton2.waitForExistence(timeout: 5) {
-                allowButton2.tap()
-            }
-        }
-
-        let openButton = safari.buttons["OPEN"]
-        if openButton.waitForExistence(timeout: 5) {
-            openButton.tap()
-        }
-    }
-}
-
-extension XCUIApplication {
-    /// Taps the specified keyboard key while handling the
-    /// keyboard onboarding interruption, if it exists.
-    /// - Parameter key: The keyboard key to tap.
-    fileprivate func tapKeyboardKey(_ key: String) {
-        let key = self.keyboards.buttons[key]
-
-        if key.isHittable == false {
-            // Attempt to find and tap the Continue button
-            // of the keyboard onboarding screen.
-            self.buttons["Continue"].tap()
-        }
-
-        key.tap()
     }
 }
